+++ beatmap/mod.ts
@@ -2,4 +2,4 @@ export * from './shared/mod.ts';
 export * as v2 from './v2/mod.ts';
 export * as v3 from './v3/mod.ts';
 export * as parse from './parse.ts';
-export * as version from './version.ts';
+export * from './version.ts';