+++ beatmap/v3/parse.ts
@@ -3,7 +3,7 @@ import { IDifficultyData } from '../../types/beatmap/v3/difficulty.ts';
 import { DifficultyData } from './difficulty.ts';
 import { DifficultyDataCheck } from './dataCheck.ts';
 import { deepCheck } from '../shared/dataCheck.ts';
-import Logger from '../../logger.ts';
+import logger from '../../logger.ts';
 
 const tag = (name: string) => {
     return `[v3::parse::${name}]`;
@@ -12,9 +12,9 @@ const tag = (name: string) => {
 const sortObjectTime = (a: IBaseObject, b: IBaseObject) => a.b - b.b;
 
 export function difficulty(data: IDifficultyData): DifficultyData {
-    Logger.info(tag('difficulty'), 'Parsing beatmap difficulty v3.x.x');
+    logger.info(tag('difficulty'), 'Parsing beatmap difficulty v3.x.x');
     if (data.version !== '3.0.0') {
-        Logger.warn(tag('difficulty'), 'Unidentified beatmap version');
+        logger.warn(tag('difficulty'), 'Unidentified beatmap version');
         data.version = '3.0.0';
     }
     deepCheck(data, DifficultyDataCheck, 'difficulty', data.version);