+++ extensions/chroma/mod.ts
@@ -1,5 +1,6 @@
 export * as types from './types/mod.ts';
 export * from './environment/mod.ts';
+export * from './lightID.ts';
 export * from './randomColor.ts';
 export * from './setColor.ts';
 export * from './shiftColor.ts';