+++ save.ts
@@ -4,8 +4,8 @@ import { DifficultyData as DifficultyDataV3 } from './beatmap/v3/difficulty.ts';
 import { ISaveOptionsDifficulty, ISaveOptionsDifficultyList, ISaveOptionsInfo } from './types/bsmap/save.ts';
 import { IDifficultyList } from './types/bsmap/list.ts';
 import { performDifficulty, performInfo } from './optimize.ts';
-import Globals from './globals.ts';
-import Logger from './logger.ts';
+import globals from './globals.ts';
+import logger from './logger.ts';
 
 const tag = (name: string) => {
     return `[save::${name}]`;
@@ -42,16 +42,16 @@ export const defaultOptions = {
  */
 export async function info(data: IInfoData, options: Partial<ISaveOptionsInfo> = {}) {
     const opt: Required<ISaveOptionsDifficulty> = {
-        path: options.path ?? (Globals.path || defaultOptions.info.path),
+        path: options.path ?? (globals.path || defaultOptions.info.path),
         filePath: options.filePath ?? 'Info.dat',
         optimise: options.optimise ?? { enabled: true },
     };
-    Logger.info(tag('info'), `Async saving info`);
+    logger.info(tag('info'), `Async saving info`);
     if (opt.optimise.enabled) {
-        Logger.info(tag('info'), `Optimising info data`);
+        logger.info(tag('info'), `Optimising info data`);
         performInfo(data, opt.optimise);
     }
-    Logger.info(tag('info'), `Writing to ${opt.path + opt.filePath}`);
+    logger.info(tag('info'), `Writing to ${opt.path + opt.filePath}`);
     await Deno.writeTextFile(opt.path + opt.filePath, JSON.stringify(data));
 }
 
@@ -62,16 +62,16 @@ export async function info(data: IInfoData, options: Partial<ISaveOptionsInfo> =
  */
 export function infoSync(data: IInfoData, options: Partial<ISaveOptionsInfo> = {}) {
     const opt: Required<ISaveOptionsDifficulty> = {
-        path: options.path ?? (Globals.path || defaultOptions.info.path),
+        path: options.path ?? (globals.path || defaultOptions.info.path),
         filePath: options.filePath ?? 'Info.dat',
         optimise: options.optimise ?? { enabled: true },
     };
-    Logger.info(tag('infoSync'), `Sync saving info`);
+    logger.info(tag('infoSync'), `Sync saving info`);
     if (opt.optimise.enabled) {
-        Logger.info(tag('infoSync'), `Optimising info data`);
+        logger.info(tag('infoSync'), `Optimising info data`);
         performInfo(data, opt.optimise);
     }
-    Logger.info(tag('infoSync'), `Writing to ${opt.path + opt.filePath}`);
+    logger.info(tag('infoSync'), `Writing to ${opt.path + opt.filePath}`);
     Deno.writeTextFileSync(opt.path + opt.filePath, JSON.stringify(data));
 }
 
@@ -85,16 +85,16 @@ export async function difficulty(
     options: Partial<ISaveOptionsDifficulty> = {},
 ) {
     const opt: Required<ISaveOptionsDifficulty> = {
-        path: options.path ?? (Globals.path || defaultOptions.difficulty.path),
+        path: options.path ?? (globals.path || defaultOptions.difficulty.path),
         filePath: options.filePath ?? data.fileName ?? 'UnnamedDifficulty.dat',
         optimise: options.optimise ?? { enabled: true },
     };
-    Logger.info(tag('difficulty'), `Async saving difficulty`);
+    logger.info(tag('difficulty'), `Async saving difficulty`);
     const objectData = data.toObject();
     if (opt.optimise.enabled) {
         performDifficulty(objectData, opt.optimise);
     }
-    Logger.info(tag('difficulty'), `Writing to ${opt.path + opt.filePath}`);
+    logger.info(tag('difficulty'), `Writing to ${opt.path + opt.filePath}`);
     await Deno.writeTextFile(opt.path + opt.filePath, JSON.stringify(objectData));
 }
 
@@ -108,16 +108,16 @@ export function difficultySync(
     options: Partial<ISaveOptionsDifficulty> = {},
 ) {
     const opt: Required<ISaveOptionsDifficulty> = {
-        path: options.path ?? (Globals.path || defaultOptions.difficulty.path),
+        path: options.path ?? (globals.path || defaultOptions.difficulty.path),
         filePath: options.filePath ?? data.fileName ?? 'UnnamedDifficulty.dat',
         optimise: options.optimise ?? { enabled: true },
     };
-    Logger.info(tag('difficultySync'), `Sync saving difficulty`);
+    logger.info(tag('difficultySync'), `Sync saving difficulty`);
     const objectData = data.toObject();
     if (opt.optimise.enabled) {
         performDifficulty(objectData, opt.optimise);
     }
-    Logger.info(tag('difficultySync'), `Writing to ${opt.path + opt.filePath}`);
+    logger.info(tag('difficultySync'), `Writing to ${opt.path + opt.filePath}`);
     Deno.writeTextFileSync(opt.path + opt.filePath, JSON.stringify(objectData));
 }
 
@@ -127,18 +127,18 @@ export function difficultySync(
  * ```
  */
 export function difficultyList(difficulties: IDifficultyList, options: Partial<ISaveOptionsDifficultyList> = {}) {
-    Logger.info(tag('difficultyList'), `Async saving list of difficulty`);
+    logger.info(tag('difficultyList'), `Async saving list of difficulty`);
     difficulties.forEach(async (dl) => {
         const opt: Required<ISaveOptionsDifficultyList> = {
-            path: options.path ?? (Globals.path || defaultOptions.difficultyList.path),
+            path: options.path ?? (globals.path || defaultOptions.difficultyList.path),
             optimise: options.optimise ?? { enabled: true },
         };
-        Logger.info(tag('difficultyListSync'), `Saving ${dl.characteristic} ${dl.difficulty}`);
+        logger.info(tag('difficultyListSync'), `Saving ${dl.characteristic} ${dl.difficulty}`);
         const objectData = dl.data.toObject();
         if (opt.optimise.enabled) {
             performDifficulty(objectData, opt.optimise);
         }
-        Logger.info(tag('difficultyList'), `Writing to ${opt.path + dl.settings._beatmapFilename}`);
+        logger.info(tag('difficultyList'), `Writing to ${opt.path + dl.settings._beatmapFilename}`);
         await Deno.writeTextFile(opt.path + dl.settings._beatmapFilename, JSON.stringify(objectData));
     });
 }
@@ -149,18 +149,18 @@ export function difficultyList(difficulties: IDifficultyList, options: Partial<I
  * ```
  */
 export function difficultyListSync(difficulties: IDifficultyList, options: Partial<ISaveOptionsDifficultyList> = {}) {
-    Logger.info(tag('difficultyListSync'), `Sync saving list of difficulty`);
+    logger.info(tag('difficultyListSync'), `Sync saving list of difficulty`);
     difficulties.forEach((dl) => {
         const opt: Required<ISaveOptionsDifficultyList> = {
-            path: options.path ?? (Globals.path || defaultOptions.difficultyList.path),
+            path: options.path ?? (globals.path || defaultOptions.difficultyList.path),
             optimise: options.optimise ?? { enabled: true },
         };
-        Logger.info(tag('difficultyListSync'), `Saving ${dl.characteristic} ${dl.difficulty}`);
+        logger.info(tag('difficultyListSync'), `Saving ${dl.characteristic} ${dl.difficulty}`);
         const objectData = dl.data.toObject();
         if (opt.optimise.enabled) {
             performDifficulty(objectData, opt.optimise);
         }
-        Logger.info(tag('difficultyList'), `Writing to ${opt.path + dl.settings._beatmapFilename}`);
+        logger.info(tag('difficultyList'), `Writing to ${opt.path + dl.settings._beatmapFilename}`);
         Deno.writeTextFileSync(opt.path + dl.settings._beatmapFilename, JSON.stringify(objectData));
     });
 }